requirements.txt
aniso8601==3.0.0
certifi==2018.4.16
click==6.7
Flask==1.0.2
flask-restplus==0.10.1
itsdangerous==0.24
Jinja2==2.10
jsonschema==2.6.0
MarkupSafe==1.0
pytz==2018.4
PyYAML==3.12
six==1.11.0
Werkzeug==0.14.1
